+++ frontend/src/common/public/index.js
@@ -1,0 +1,6 @@
+import HospitalIcon from './HospitalIcon'
+
+
+export {
+  HospitalIcon
+}
