+++ frontend/src/common/views/fields/SimpleView.jsx
@@ -3,6 +3,7 @@ import React from 'react';
 import SimpleViewInline from './inline/SimpleViewInline';
 import SimpleViewRegular from './regular/SimpleViewRegular';
 
+
 const SimpleView = ({
   title, description,
   isInline,
