+++ frontend/src/common/views/fields/inline/ChoiceViewInline.jsx
@@ -1,11 +1,13 @@
+import React from 'react';
+
 import {
   Button,
   Col,
   Container, Form, Row,
 } from 'react-bootstrap';
-import React from 'react';
 import BootstrapSwitchButton from 'bootstrap-switch-button-react';
 import InlineView from './common/InlineView';
+
 
 const ChoiceFormTwoValues = ({
   title, choices,
