+++ frontend/src/common/views/fields/inline/SimpleViewInline.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
+
 import InlineView from './common/InlineView';
+
 
 const SimpleViewInline = ({
   title,
