+++ frontend/src/common/views/fields/inline/common/InlineView.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 
+
 const InlineView = ({
   children, highlighted,
 }) => (
