+++ frontend/src/common/views/fields/regular/ChoiceViewRegular.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
+
 import { Form } from 'react-bootstrap';
 import SingleInput from './common/SingleInput';
+
 
 const ChoiceForm = ({
   choices,
