+++ frontend/src/common/views/fields/regular/common/SingleInputButton.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
+
 import { Button, Col, Row } from 'react-bootstrap';
 import { FaEye, FaEyeSlash } from 'react-icons/fa';
+
 
 const Header = ({
   message,
