+++ frontend/src/common/views/index.js
@@ -1,0 +1,20 @@
+import ChoiceView from './fields/ChoiceView'
+import DerivedView from './fields/DerivedView'
+import SimpleView from './fields/SimpleView'
+import SliderView from './fields/SliderView'
+import TextView from './fields/TextView'
+
+import SignatureView from './signature/SignatureView'
+
+import LoadingView from './utility/LoadingView'
+
+
+export {
+  ChoiceView,
+  DerivedView,
+  SimpleView,
+  SliderView,
+  TextView,
+  SignatureView,
+  LoadingView
+}
