+++ frontend/src/common/views/signature/SignatureView.jsx
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
+
+import { Button, Row , Container} from 'react-bootstrap';
 import SignatureCanvas from 'react-signature-canvas';
-import { Button, Row } from 'react-bootstrap';
-import Container from 'react-bootstrap/Container';
+
 
 const Header = ({ message }) => (
   <Row>
